+++ src/routes/login/components/LoginForm.tsx
@@ -16,6 +16,8 @@ import {
     AlertIcon,
     AlertTitle,
     AlertDescription,
+    Link,
+    Divider,
 } from '@chakra-ui/react';
 import { RiErrorWarningLine, RiEyeLine, RiEyeOffLine } from 'react-icons/ri';
 
@@ -32,6 +34,7 @@ const LoginForm = (): JSX.Element => {
     const [showPassword, setShowPassword] = useState(false);
     const dispatch = useAppDispatch();
     const [, setLocation] = useLocation();
+    const authInfo = useAppSelector((state) => state.me.info.auth);
 
     const organizationId = useAppSelector(
         (state) => state.me.info.organization_id
@@ -39,15 +42,14 @@ const LoginForm = (): JSX.Element => {
     const userLoading = useAppSelector((state) => state.me.loading);
     const userError = useAppSelector((state) => state.me.error);
 
+    const urlSearchParams = new URLSearchParams(window.location.search);
+    const nextLocation = urlSearchParams.get('next') || PATHS.COMPUTE_PLANS;
+
     const submitLogin = async (username: string, password: string) => {
         const payload: LoginPayloadT = {
             username,
             password,
         };
-        const urlSearchParams = new URLSearchParams(window.location.search);
-
-        const nextLocation = urlSearchParams.get('next') || PATHS.COMPUTE_PLANS;
-
         dispatch(logIn(payload))
             .then(unwrapResult)
             .then(
@@ -75,6 +77,25 @@ const LoginForm = (): JSX.Element => {
                 <Text fontWeight="semibold" fontSize="3xl" marginBottom="4">
                     Login to {organizationId}
                 </Text>
+
+                {authInfo?.oidc && (
+                    <Box>
+                        <Button
+                            width="100%"
+                            marginBottom="4"
+                            as={Link}
+                            href={`${API_URL}${
+                                authInfo.oidc.login_url
+                            }?next=${encodeURIComponent(
+                                window.location.origin + nextLocation
+                            )}`}
+                        >
+                            Sign in with {authInfo.oidc.name}
+                        </Button>
+                        <Divider marginBottom="4" />
+                    </Box>
+                )}
+
                 {userError && (
                     <Alert
                         marginY="4"