+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Inter font for content body (#176)
 -   Task duration section in Task Drawer (#165)
+-   SSO on login page and API keys modal (#181)
 
 ### Changed